requirements.txt
azure_storage==0.37.0
openai==0.27.2
pydub==0.25.1
python-dotenv==1.0.0
